+++ src/api/index.ts
@@ -251,7 +251,7 @@ export enum TerminalWidthProfile {
 // mjb_next_line_state
 // mjb_next_word_state
 // mjb_next_sentence_state
-// mjb_for_each_character_fn
+// mjb_for_each_codepoint_fn
 
 // mjb_direction
 export enum Direction {
@@ -307,7 +307,7 @@ export enum IdentifierProfile {
   NFKC,
 };
 
-// Result of mjb_for_each_character function
+// Result of mjb_for_each_codepoint function
 export type NextCharacter = {
   character: Character;
   type: NextCharacterType; // mjb_character_position
@@ -597,15 +597,15 @@ export class Mojibake {
     }
   }
 
-  // mjb_status mjb_for_each_character(const char *buffer, size_t byte_length,
-  // mjb_encoding encoding, mjb_for_each_character_fn callback)
-  forEachCharacter(input: MojibakeInput, options: TextInputOptions = {}): NextCharacter[] | null {
+  // mjb_status mjb_for_each_codepoint(const char *buffer, size_t byte_length,
+  // mjb_encoding encoding, mjb_for_each_codepoint_fn callback)
+  forEachCodepoint(input: MojibakeInput, options: TextInputOptions = {}): NextCharacter[] | null {
     const wasmInput = this.copyInput(input, options.encoding);
-    const previousCallback = (globalThis as any)._mjbForEachCharacterCallback;
+    const previousCallback = (globalThis as any)._mjbForEachCodepointCallback;
     const characters: NextCharacter[] = [];
 
-    // See mjb_for_each_character function
-    (globalThis as any)._mjbForEachCharacterCallback = (character: Pointer, type: number) => {
+    // See mjb_for_each_codepoint function
+    (globalThis as any)._mjbForEachCodepointCallback = (character: Pointer, type: number) => {
       characters.push({
         character: this.pointerToCharacter(character),
         type
@@ -615,7 +615,7 @@ export class Mojibake {
     };
 
     try {
-      const status = this.module._mjb_for_each_character(wasmInput.ptr, wasmInput.size,
+      const status = this.module._mjb_for_each_codepoint(wasmInput.ptr, wasmInput.size,
         wasmInput.encoding, 0);
 
       if(status !== Status.OK) {
@@ -625,9 +625,9 @@ export class Mojibake {
       return characters;
     } finally {
       if(previousCallback === undefined) {
-        delete (globalThis as any)._mjbForEachCharacterCallback;
+        delete (globalThis as any)._mjbForEachCodepointCallback;
       } else {
-        (globalThis as any)._mjbForEachCharacterCallback = previousCallback;
+        (globalThis as any)._mjbForEachCodepointCallback = previousCallback;
       }
 
       this.free(wasmInput.ptr);