+++ src/api/mojibake.d.ts
@@ -45,8 +45,8 @@ export type MojibakeWasmModule = {
   _mjb_is_utf16: (buffer: Pointer, byte_length: number) => boolean;
   // Return the length of a string.
   _mjb_count_codepoints: (buffer: Pointer, max_length: number, encoding: number) => number;
-  // Run a callback for each character of a string.
-  _mjb_for_each_character: (buffer: Pointer, byte_length: number, encoding: number, callback: number) => number;
+  // Run a callback for each codepoint of a string.
+  _mjb_for_each_codepoint: (buffer: Pointer, byte_length: number, encoding: number, callback: number) => number;
   // Return the value of a binary Unicode property.
   _mjb_codepoint_property_binary: (codepoint: Codepoint, property: number, value: Pointer) => number;
   // Return the value of an enumerated or integer Unicode property.