+++ src/shell/commands/character.c
@@ -276,7 +276,7 @@ static bool mjbsh_output_next_character(mjb_character *character, mjb_character_
 }
 
 int mjbsh_character_command(int argc, char *const argv[], unsigned int flags) {
-    mjb_status status = mjb_for_each_character(argv[0], strlen(argv[0]), MJB_ENC_UTF_8,
+    mjb_status status = mjb_for_each_codepoint(argv[0], strlen(argv[0]), MJB_ENC_UTF_8,
         mjbsh_output_next_character);
 
     if(status != MJB_STATUS_OK) {